Favicon downloader: Try to get og:image first
Some websites provide really small favicons.

Facebook for instance provides a 16x16px favicon which looks really blurry when
used on the desktop (Alt-tab, dock etc..).

It might be possible to fake the user agent and define the screen resolution
within the request to force such websites to give better resolutions (when asked
via Firefox directly, Facebook gives 32x32px for instance).

In any case, if the website supports og:image, we should try and use that. This
gives us a nice high-res icon (325x325px in the face of Facebook).